Revamp search method
The search method returns a string representation of one item based on
a single name field match. This is unintuitive and reduces code reusability.

It has been revamped to allow searching through arbitrary fields based
on the order of arguments passed in.

Let's try to call on this method for ItemList based operations,
overloading it with more specific argument fields as needed.