Extract search methods into separate class
The logic for searching through the inventory resides in the ItemList
class. This obscures it from use by other parts of the code that need
it.

By separating it into its own class, we have a common interface to build
upon other commands requiring search. In addition, the testability of
our code increases.

Let's further segment logical units of the ItemList to improve on
separation of concerns.